+++ lib/libesp32/berry_matter/src/be_matter_module.c
@@ -258,6 +258,8 @@ extern const bclass be_class_Matter_TLV;   // need to declare it upfront because
 #include "solidify/solidified_Matter_Plugin_3_Sensor_Contact.h"
 #include "solidify/solidified_Matter_Plugin_3_Sensor_Rain.h"
 #include "solidify/solidified_Matter_Plugin_3_Sensor_Waterleak.h"
+#include "solidify/solidified_Matter_Plugin_2_Fan.h"
+#include "solidify/solidified_Matter_Plugin_9_Virt_Fan.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Sensor_Contact.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Sensor_Occupancy.h"
 #include "solidify/solidified_Matter_Plugin_9_Virt_Sensor_Rain.h"