+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Extend command ``SetOption147 1`` to disable publish of IRReceived MQTT messages (#21574)
 - Matter support for Rain sensor (#21633)
 - Matter add internal debug option
+- Matter add Fan support (virtual only)
 
 ### Breaking Changed